test: move balance tests to shared suite
Move 'internal transfer' and 'custom token' tests
from fullnode-specific to shared get-balance suite.
Extract fullnode-only assertions (nonexistent token
zero-balance, cross-wallet isolation) into dedicated
fullnode-specific tests.
